Bot/requirements.txt
python-telegram-bot==21.0.1
requests==2.31.0
python-dotenv==1.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0